backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0
gunicorn>=21.0,<22.0
whitenoise>=6.6,<7.0
django-filter>=23.5,<24.0
